+++ ChangeLog.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+- NEW : Ajout de la class TechATM pour l'affichage de la page "A propos" *11/05/2022* 1.8.0
+
 ## 1.7
 - FIX: date facturation n'était jamais mise à jour - *2021-12-09* - 1.7.3
 - FIX: compliance with dolistore rules for `main.inc.php` inclusion - *2021-10-20* - 1.7.2